+++ src/components/Layout/index.tsx
@@ -1,11 +1,38 @@
-import { Outlet } from 'react-router-dom';
+import { useCallback, useState } from 'react';
+import { Outlet, useLocation, useNavigate } from 'react-router-dom';
+import { useAppDispatch } from '../../store/hook';
+import { getTokenWithLocalStorage } from '../../store/signInUpSlice';
 import { Footer } from '../Footer';
+import { Header } from '../Header';
+
+import styles from './styles.module.scss';
 
 export const Layout = () => {
+  const location = useLocation();
+  const [isCheckedTheme, setIsCheckedTheme] = useState(true);
+
+  const navigate = useNavigate();
+
+  const dispatch = useAppDispatch();
+
+  const logOut = useCallback(() => {
+    localStorage.clear();
+
+    dispatch(getTokenWithLocalStorage(''));
+    navigate('/');
+  }, [navigate, dispatch]);
+
   return (
-    <>
-      <Outlet />
-      <Footer />
-    </>
+    <section className={isCheckedTheme ? styles.darkTheme : styles.lightTheme}>
+      <header>{location.pathname === '/main' && <Header signOut={logOut} />}</header>
+
+      <main>
+        <Outlet />
+      </main>
+
+      <footer className={styles.footerWrapper}>
+        <Footer />
+      </footer>
+    </section>
   );
 };
