+++ src/components/pages/MainPage/index.tsx
@@ -7,20 +7,8 @@ import { Header } from '../../Header';
 import { Main } from '../../Design/Main';
 
 export const MainPage = () => {
-  const navigate = useNavigate();
-
-  const dispatch = useAppDispatch();
-
-  const logOut = useCallback(() => {
-    localStorage.clear();
-
-    dispatch(getTokenWithLocalStorage(''));
-    navigate('/');
-  }, [navigate, dispatch]);
-
   return (
     <div>
-      <Header signOut={logOut} />
       <Main />
     </div>
   );
